- Merged the original PhpUnit plugin with the rewrite
